Выделяет компонет popup-controls
