+++ src/components/comments.js
@@ -36,13 +36,12 @@ const cteateCommentTemplate = (comments) => {
   const count = comments.length;
 
   return (`
-  <div class="form-details__bottom-container">
+
   <section class="film-details__comments-wrap">
     <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${count}</span></h3>
     <ul class="film-details__comments-list">
     ${commentsMarkup}
     </ul>
-
     <div class="film-details__new-comment">
       <div for="add-emoji" class="film-details__add-emoji-label"></div>
       <label class="film-details__comment-label">
@@ -53,7 +52,7 @@ const cteateCommentTemplate = (comments) => {
       </div>
     </div>
   </section>
-  </div>
+
 `);
 };
 export {cteateCommentTemplate};
